+++ model/seat.go
@@ -1,7 +1,7 @@
 package model
 
 type Seat struct {
-	ID          int64  `json:"id" gorm:"primaryKey; autoIncrement"`
+	ID          int    `json:"id" gorm:"primaryKey; autoIncrement"`
 	RoomID      int    `json:"room_id" gorm:"index; not null"`
 	Name        string `json:"name"`
 	Description string `json:"description"`
